+++ CHANGELOG.md
@@ -10,9 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Web UI for chatting with LLM.
-- Ollama and Anthropic LLM providers.
+- Ollama, Anthropic, OpenAI and OpenRouter LLM providers.
 - Configuration file for LLM provider.
 - Bolt DB for chat and message persistence.
 - List MCP objects (servers, tools, resources, prompts) in the UI.
 - Customizable system and title generation prompts.
 - Implement tools interaction for anthropic.
+- Own LLM for title generation.